python-codecs: restructure, add gstreamer decoder option
