+++ plugins/python-codecs/package.json
@@ -1,6 +1,6 @@
 {
    "name": "@scrypted/python-codecs",
-   "version": "0.1.6",
+   "version": "0.1.7",
    "description": "Python Codecs for Scrypted",
    "keywords": [
       "scrypted",
